added UMD pattern and grunt to build project
