+++ demos/debug/index.html
@@ -9,11 +9,11 @@
 <meta name="mobile-web-app-capable" content="yes">
 
 <meta name="apple-mobile-web-app-title" content="Add to Home">
-<link rel="shortcut icon" sizes="16x16" href="../../imgs/icon-16x16.png">
-<link rel="shortcut icon" sizes="196x196" href="../../imgs/icon-196x196.png">
-<link rel="apple-touch-icon-precomposed" href="../../imgs/icon-152x152.png">
+<link rel="shortcut icon" sizes="16x16" href="../../dist/img/icon-16x16.png">
+<link rel="shortcut icon" sizes="196x196" href="../../dist/img/icon-196x196.png">
+<link rel="apple-touch-icon-precomposed" href="../../dist/img/icon-152x152.png">
 
-<link rel="stylesheet" type="text/css" href="../../style/addtohomescreen.css">
+<link rel="stylesheet" type="text/css" href="../../dist/style/addtohomescreen.css">
 </head>
 
 <body>
@@ -21,7 +21,7 @@
 
 <p>You can clear the user session with <strong>addToHomescreen.removeSession()</strong> or <strong>instance.clearSession()</strong>. The first completely removes the localStorage variable, the second simply resets the session to the defaults.</p>
 
-<script src="../../src/addtohomescreen.js"></script>
+<script src="../../dist/addtohomescreen.min.js"></script>
 <script>
 // addToHomescreen.removeSession();     // use this to remove the localStorage variable
 var ath = addToHomescreen({
