+++ demos/detect-homescreen/index.html
@@ -6,12 +6,12 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 
 <meta name="apple-mobile-web-app-title" content="Add to Home">
-<link rel="shortcut icon" sizes="16x16" href="../../imgs/icon-16x16.png">
-<link rel="shortcut icon" sizes="196x196" href="../../imgs/icon-196x196.png">
-<link rel="apple-touch-icon-precomposed" href="../../imgs/icon-152x152.png">
+<link rel="shortcut icon" sizes="16x16" href="../../dist/img/icon-16x16.png">
+<link rel="shortcut icon" sizes="196x196" href="../../dist/img/icon-196x196.png">
+<link rel="apple-touch-icon-precomposed" href="../../dist/img/icon-152x152.png">
 
-<link rel="stylesheet" type="text/css" href="../../style/addtohomescreen.css">
-<script src="../../src/addtohomescreen.js"></script>
+<link rel="stylesheet" type="text/css" href="../../dist/style/addtohomescreen.css">
+<script src="../../dist/addtohomescreen.min.js"></script>
 <script>
 addToHomescreen({
     detectHomescreen: true
