+++ demos/simple/index.html
@@ -10,18 +10,18 @@
 
 <meta name="apple-mobile-web-app-title" content="Add to Home">
 
-<link rel="shortcut icon" sizes="16x16" href="../../imgs/icon-16x16.png">
-<link rel="shortcut icon" sizes="196x196" href="../../imgs/icon-196x196.png">
+<link rel="shortcut icon" sizes="16x16" href="../../dist/icon-16x16.png">
+<link rel="shortcut icon" sizes="196x196" href="../../dist/icon-196x196.png">
 <!--link rel="apple-touch-icon-precomposed" sizes="152x152" href="icon-152x152.png">
 <link rel="apple-touch-icon-precomposed" sizes="144x144" href="icon-144x144.png">
 <link rel="apple-touch-icon-precomposed" sizes="120x120" href="icon-120x120.png">
 <link rel="apple-touch-icon-precomposed" sizes="114x114" href="icon-114x114.png">
 <link rel="apple-touch-icon-precomposed" sizes="76x76" href="icon-76x76.png">
 <link rel="apple-touch-icon-precomposed" sizes="72x72" href="icon-72x72.png"-->
-<link rel="apple-touch-icon-precomposed" href="../../imgs/icon-152x152.png">
+<link rel="apple-touch-icon-precomposed" href="../../dist/icon-152x152.png">
 
-<link rel="stylesheet" type="text/css" href="../../style/addtohomescreen.css">
-<script src="../../src/addtohomescreen.js"></script>
+<link rel="stylesheet" type="text/css" href="../../dist/style/addtohomescreen.css">
+<script src="../../dist/addtohomescreen.min.js"></script>
 <script>
 addToHomescreen();
 </script>
